Add files via upload
Lab 2 is a two-step practice that involves working with text files and Oracle cloud. In the first step, I download a large text file and write a Python code to count the number of words in the file.

In the second step, I set up the Oracle Cloud Infrastructure (OCI) and use Python SDK. I create an object storage bucket in OCI and upload the text file to a specific folder within the bucket. I write a Python code that can count the words in the file when executed on a DataFlow application in OCI.